% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.addAspectNameToJson}
\alias{.addAspectNameToJson}
\title{Add the aspect name to the JSON}
\usage{
.addAspectNameToJson(json, name)
}
\arguments{
\item{json}{character; preformated json}

\item{name}{character; name of the aspect}
}
\value{
character; character of json object
}
\description{
Add the aspect name to the JSON
}
\note{
Internal function only for convenience
}
\examples{
json = '{bla:"BLA", blubb:"BLUBB"}'
RCX:::.addAspectNameToJson(json, "foo")
}
\keyword{internal}
