% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igraph.R
\name{.addAttributeData}
\alias{.addAttributeData}
\title{Add attribute data to an igraph object}
\usage{
.addAttributeData(ig, attributeRef, attribute)
}
\arguments{
\item{ig}{igraph object}

\item{attributeRef}{reference name; "node", "edge" or "network"}

\item{attribute}{an attribute aspect}
}
\value{
igraph object
}
\description{
Not only simply add the name-value pairs, but also:
\itemize{
\item unlist lists if indicated by isList column
\item renames name="name" to "attribute$name"
\item puts subnetwork id at the and of the attribute name
\item adds a data type as attribute$dataType if not string, boolean or double
}
}
\note{
Internal function only for convenience
}
\examples{
NULL
}
\keyword{internal}
