% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-general.R
\name{.aspectClass}
\alias{.aspectClass}
\title{Get the class of aspects}
\usage{
.aspectClass(x)
}
\arguments{
\item{x}{a potential aspect}
}
\value{
The aspect class name or \code{NA} if it's not an aspect
}
\description{
Get the class of aspects
}
\note{
Internal function only for convenience
}
\examples{
x = list(a="foo", b="bar")
class(x)
# [1] "list"
\dontrun{
.addClass(x) <- .CLS$nodes
class(x)
# [1] "NodesAspect" "list" 

.aspectClass(x)
# [1] "NodesAspect"

.removeClass(x) <- "NodesAspect"

.aspectClass(x)
# [1] "NA"

}
}
\keyword{Only}
\keyword{`aspectClasses`}
\keyword{are}
\keyword{aspects}
\keyword{class}
\keyword{defined}
\keyword{get}
\keyword{in}
\keyword{internal}
\keyword{of}
\keyword{that}
\keyword{the}
