% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.convertRawList}
\alias{.convertRawList}
\title{Convert a list of vectors to a character vector with pasted elements}
\usage{
.convertRawList(l, keepNa = T)
}
\arguments{
\item{l}{unnamed list}

\item{keepNa}{logical; whether to keep \code{NA} values or replace it with an empty list}
}
\value{
character
}
\description{
Convert a list of vectors to a character vector with pasted elements
}
\note{
Internal function only for convenience
}
\examples{
l  = list(NA,c(2,3), 5)
RCX:::.convertRawList(l)
}
\keyword{internal}
