% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-general.R
\name{.createAttributeAspect}
\alias{.createAttributeAspect}
\title{Create a default *AttributeAspect}
\usage{
.createAttributeAspect(
  propertyOf,
  name,
  value,
  dataType,
  isList,
  subnetworkId = NULL,
  .log = ""
)
}
\arguments{
\item{propertyOf}{integer; refers to the IDs of an other aspect}

\item{name}{character; key of the attribute}

\item{value}{character; value of the attribute}

\item{dataType}{character (optional);}

\item{isList}{logical (optional);}

\item{subnetworkId}{integer (optional); CySubNetworks}
}
\value{
*AttributeAspect prototype object
}
\description{
Some aspects like \emph{NodeAttributesAspect} or \emph{EdgeAttributesAspect} use a key-value scheme.
This function helps in constructing while avoiding repetition.
}
\note{
Internal function only for convenience
}
\examples{
NULL
}
\seealso{
\code{\link{.mergeIdAspect}}, \code{\link{.mergeAttributesAspect}}, \code{\link{CySubNetworks}}
}
\keyword{internal}
