% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.filterBy}
\alias{.filterBy}
\title{Filter several parameters for elements, that match to a given name in a given param}
\usage{
.filterBy(name, param, ...)
}
\arguments{
\item{name}{character; matching value}

\item{param}{character; in which param in ...}

\item{...}{several parameters}
}
\value{
list with only matching elements of all parameters
}
\description{
Filter several parameters for elements, that match to a given name in a given param
}
\note{
Internal function only for convenience
}
\examples{
po=c("match","not","some","other","match")
prop=c(1,2,3,4,5)
dep=c("bla","blubb","bla","blubb","bla")
map=list("BLA","BLUBB","BLA","BLUBB","BLA") 

RCX:::.filterBy("match", "po", po, prop, dep, map)
}
\keyword{internal}
