% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.json2RDataType}
\alias{.json2RDataType}
\title{Get the data type and isList from JSON data}
\usage{
.json2RDataType(dataType, default = "string")
}
\arguments{
\item{dataType}{data type column from jsonToRCX => .jsonV}

\item{default}{default value for \code{NA} values (by default the values remain \code{NA})}
}
\value{
\verb{list(type=<character vector>, isList=<logical vector>)}
}
\description{
Get the data type and isList from JSON data
}
\note{
Internal function only for convenience
}
\examples{
jsonD = c("boolean", "double", "integer", "long", "string", 
          "list_of_boolean", "list_of_double", "list_of_integer", 
          "list_of_long", "list_of_string")

RCX:::.json2RDataType(jsonD)
}
\keyword{internal}
