% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.jsonV}
\alias{.jsonV}
\title{Return data as a vector from a JSON list}
\usage{
.jsonV(data, acc, default = NA, returnAllDefault = TRUE)
}
\arguments{
\item{data}{json list}

\item{acc}{accession name}

\item{default}{default return value}

\item{returnAllDefault}{whether to return the vector if all values are the default value (or \code{NULL} instead)}
}
\value{
vector
}
\description{
Return data as a vector from a JSON list
}
\note{
Internal function only for convenience
}
\examples{
testData = list(list(n="CDKN1B"),
                list(n="ROCK1", r="BLA"),
                list(n="SHC1", r="BLUBB"),
                list(n="IRS1"))
RCX:::.jsonV(testData, "r")
}
\keyword{internal}
