% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-general.R
\name{.modClass}
\alias{.modClass}
\alias{.addClass<-}
\alias{.addClass}
\alias{.removeClass<-}
\alias{.removeClass}
\title{Add or remove a class name from an object}
\usage{
.addClass(x) <- value

.removeClass(x) <- value
}
\arguments{
\item{x}{an R object.}

\item{value}{character vector of lenght 1.}
}
\value{
an R object.
}
\description{
Add or remove a class name from an object
}
\note{
Internal function only for convenience
}
\examples{
x = list(a="foo", b="bar")
class(x)
# [1] "list"
\dontrun{
.addClass(x) <- "blubb"
class(x)
# [1] "blubb" "list" 

.addClass(x) <- "blubb"
class(x)
# [1] "blubb" "list" 

.removeClass(x) <- "blubb"
class(x)
# [1] "list"
}
}
\keyword{internal}
