% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-general.R
\name{.pasteC}
\alias{.pasteC}
\title{Concatenate as comma separated character vector}
\usage{
.pasteC(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
character vector of length 1.
}
\description{
Concatenate as comma separated character vector
}
\note{
Internal function only for convenience
}
\examples{
\dontrun{
a <- c("one", "two", "three")
.pasteC(a)
#[1] "one, two, three"
}
}
\keyword{internal}
