% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{.renameDF}
\alias{.renameDF}
\title{Rename data.frame columns by key-value pairs in rnames}
\usage{
.renameDF(df, rnames)
}
\arguments{
\item{df}{data.frame}

\item{rnames}{named character vector; \code{names(rnames)=colnames(df)}}
}
\value{
df with new colnames; or NULL on error
}
\description{
Rename data.frame columns by key-value pairs in rnames
}
\note{
Internal function only for convenience
}
\examples{
nodes = data.frame(id=c(0,1,2),
                   name=c("CDK1",NA,"CDK3"),
                   represents=c(NA,"bla",NA))
rnames = c(id="@id", name="n", represents="r")
RCX:::.renameDF(nodes, rnames)
}
\keyword{internal}
