% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{.summaryAspect}
\alias{.summaryAspect}
\title{Helper function to mark columns that are ids or reference ids}
\usage{
.summaryAspect(aspect)
}
\arguments{
\item{aspect}{an aspect (data.frame)}
}
\value{
the aspect (data.frame)
}
\description{
Helper function to mark columns that are ids or reference ids
}
\note{
Internal function only for convenience
}
\examples{
edges = createEdges(source=c(1,1), target=c(2,3))
edges = RCX:::.summaryAspect(edges)
class(edges$id)
}
\keyword{internal}
