% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{.transformListLength<-}
\alias{.transformListLength<-}
\title{Transform an aspect with a list as column}
\usage{
.transformListLength(aspect) <- value
}
\arguments{
\item{aspect}{an aspect (data.frame)}

\item{value}{character; property}
}
\value{
the aspect (data.frame)
}
\description{
Transforms an aspect with a column, that is a list to force a custom format
in summary generation. Only show the number of elements in the list elements.
}
\note{
Internal function only for convenience
}
\examples{
df = data.frame(bla=c("a","b","c"))
df$blubb=list(c("a","b","c"),
              c(1,2),
              c(TRUE,FALSE,FALSE,TRUE,TRUE))

RCX:::.transformListLength(df) = "blubb"

summary(df)
}
\keyword{internal}
