% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{.transformVLD}
\alias{.transformVLD}
\title{Transform an aspect with data type}
\usage{
.transformVLD(aspect)
}
\arguments{
\item{aspect}{an aspect (data.frame)}
}
\value{
the aspect (data.frame)
}
\description{
Transforms an aspect with \code{value}, \code{dataType} and \code{isList} columns to force a custom format
in summary generation.
}
\note{
Internal function only for convenience
}
\examples{
df = data.frame(bla=c("a","b","c"),
                value=list("a",2,TRUE),
                dataType=c("string","integer","boolean"),
                isList=c(FALSE,FALSE,FALSE))
df = RCX:::.transformVLD(df)

summary(df)
}
\keyword{internal}
