% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{.validateAttributesAspect}
\alias{.validateAttributesAspect}
\title{Helper for validating node- and edge-attributes aspects}
\usage{
.validateAttributesAspect(aspect, verbose = TRUE)
}
\arguments{
\item{aspect}{an RCX aspect}

\item{verbose}{logical; whether to print the test results.}
}
\value{
logical; whether the test passed
}
\description{
Helper for validating node- and edge-attributes aspects
}
\note{
Internal function only for convenience
}
\keyword{internal}
