% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json-fromJson.R
\name{fromCX}
\alias{fromCX}
\title{Convert CX to an \link[=RCX-object]{RCX} object}
\usage{
fromCX(cx, verbose = F, raw = F)
}
\arguments{
\item{cx}{CX file in JSON format}

\item{verbose}{logical; whether to print what is happening}

\item{raw}{logical; whether to return unparsable aspects}
}
\value{
\link[=RCX-object]{RCX} object; (or a \verb{list(rcx=<}\link[=RCX-object]{RCX}\verb{>, raw=<raw>)} if \code{raw=TRUE})
}
\description{
This function parses a CX file or text and creates an \link[=RCX-object]{RCX} object from it.
}
\details{
The CX-JSON is parsed using the \link{jsonlite} package, with the \link{jsonToRCX} functions creating objects for the single aspects.
For standard CX aspects are processed by generic functions named by the aspect names of the CX data structure
(see NDEx documentation: \url{https://home.ndexbio.org/data-model/}).

The CX network may contain additional aspects besides the officially defined ones.
This includes self defined or deprecated aspects, that sill can be found in the networks at the NDEx platform.
By default, those aspects are simply omitted.

In those cases, the setting \emph{verbose} to \code{TRUE} is a good idea to see, which aspects cannot be processed this package.
If \emph{raw} is set to \code{TRUE}, those aspects are also returned along the \link[=RCX-object]{RCX} object in \strong{raw} format,
i.e. as a named list containing the parsed JSON from

\code{jsonlite::fromJSON(cx, simplifyVector = F)}

The raw data can be handled individually, but is advisable to extend the \link{jsonToRCX} functions by implementing own versions
for those aspects.

Additionally, the \strong{update} functions have to be implemented to add the newly generated aspect objects
to \link[=RCX-object]{RCX} object (see e.g. \link{updateNodes} or \link{updateEdges}).
Therefore, the function also have to be named \verb{"update<aspect-name>}, where aspect-name is the capitalized version of the
name used in the CX.
}
\examples{
NULL
}
\seealso{
\link{toCX}, \link{rcxToJson}, \link{readRCX}, \link{writeCX}
}
