% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igraph.R
\name{Igraph}
\alias{Igraph}
\alias{toIgraph}
\alias{fromIgraph}
\title{Convert an RCX object from and to an igraph object}
\usage{
toIgraph(rcx, directed = FALSE)

fromIgraph(
  ig,
  nodeId = "id",
  nodeName = "nodeName",
  nodeIgnore = c("name"),
  edgeId = "id",
  edgeIgnore = c("name"),
  suppressWarning = F
)
}
\arguments{
\item{rcx}{\link[=RCX-object]{RCX} object}

\item{directed}{logical; whether the graph is directed}

\item{ig}{\link[igraph:aaa-igraph-package]{igraph} object}

\item{nodeId}{character; igraph attribute name used for \link[=Nodes]{node} ids}

\item{nodeName}{character; igraph attribute name used for \link[=Nodes]{node} names}

\item{nodeIgnore}{character; igraph attribute names that should be ignored}

\item{edgeId}{character; igraph attribute name used for \link[=Edges]{edge} ids}

\item{edgeIgnore}{character; igraph attribute names that should be ignored}

\item{suppressWarning}{logical; whether to suppress a warning message, if the validation of the \link[=RCX-object]{RCX} object fails}
}
\value{
\link[igraph:aaa-igraph-package]{igraph} or \link[=RCX-object]{RCX} object
}
\description{
Convert an \link[=RCX-object]{RCX} object to an \link[igraph:aaa-igraph-package]{igraph} object
}
\examples{
NULL
}
\seealso{
\link{toCX}, \link{rcxToJson}, \link{readRCX}, \link{writeCX}
}
