% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{markAttributeColumn}
\alias{markAttributeColumn}
\alias{.markAttributeColumn<-}
\title{Mark attribute name columns within a data.frame}
\usage{
.markAttributeColumn(aspect) <- value
}
\arguments{
\item{aspect}{an aspect (data.frame)}

\item{value}{character; property}
}
\value{
the aspect (data.frame)
}
\description{
Assigns a class to a data.frame column to force a custom format in summary generation.
}
\note{
Internal function only for convenience
}
\examples{
df = data.frame(name=c("a","b","c"),
                value=c("a","b","c"))
RCX:::.markRefColumn(df) = "name"

summary(df)
}
\keyword{internal}
