% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{markRefColumn}
\alias{markRefColumn}
\alias{.markRefColumn<-}
\alias{.markReqRefColumn<-}
\title{Mark required and optional references within a data.frame}
\usage{
.markRefColumn(aspect) <- value

.markReqRefColumn(aspect) <- value
}
\arguments{
\item{aspect}{an aspect (data.frame)}

\item{value}{character; property}
}
\value{
the aspect (data.frame)
}
\description{
Assigns a class to a data.frame column to force a custom format in summary generation.
}
\note{
Internal function only for convenience
}
\examples{
df = data.frame(bla=c("a","b","c"),
                blubb=c("a","b","c"))
RCX:::.markRefColumn(df) = "bla"

summary(df)
}
\keyword{internal}
