% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCX.R
\docType{package}
\name{RCX-package}
\alias{RCX-package}
\alias{RCX}
\title{RCX and RCXgraph data structure package}
\description{
R package implementing the Cytoscape Exchange (CX) format
}
\details{
\tabular{ll}{
Package: \tab RCX\cr
Type: \tab Package\cr
Version: \tab 2.0.7\cr
Date: \tab 2016-12-02\cr
License: \tab TBD\cr
}
}
\examples{
\dontrun{
require(RCX)
###TODO: add some sample code
}
}
\author{
Florian Auer \email{florian.auer@informatik.uni-augsburg.de}
}
\keyword{package}
