% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_RCX.r
\name{rcx_fromFile}
\alias{rcx_fromFile}
\title{Reads an RCX object from a file in CX (JSON) format}
\usage{
rcx_fromFile(file, verbose = FALSE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a single string or a raw vector)}

\item{verbose}{logical; whether to print out extended feedback}
}
\value{
json object if successfull, empty string otherwise
}
\description{
Reads an RCX object from a file in CX (JSON) format
}
\examples{
## Read from a CX file
#rcx = rcx_fromFile('/path/to/file.cx')
}
\seealso{
\code{\link{rcx_fromJSON}}
}
