% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json-fromJson.R
\name{readRCX}
\alias{readRCX}
\alias{readCX}
\title{Read CX from file}
\usage{
readRCX(file, verbose = F, raw = F)

readCX(file, verbose = F)
}
\arguments{
\item{file}{character; the name of the file which the data are to be read from}

\item{verbose}{logical; whether to print what is happening}

\item{raw}{logical; whether to return unparsable aspects (see \link{fromCX})}
}
\description{
These function read a CX network from a file.
}
\details{
These functions return either simply the read JSON from the file, or
convert the CX to an \link[=RCX-object]{RCX} by calling \link{fromCX}.
}
\examples{
NULL
}
\seealso{
\link{fromCX}, \link{jsonToRCX}, \link{writeCX}
}
