% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json-toJson.R
\name{writeCX}
\alias{writeCX}
\alias{writeCX.RCX}
\alias{writeCX.CX}
\title{Write RCX to file}
\usage{
writeCX(x, file, verbose = F, pretty = F)

\method{writeCX}{RCX}(x, file, verbose = F, pretty = F)

\method{writeCX}{CX}(x, file, verbose = F, pretty = F)
}
\arguments{
\item{x}{\link[=RCX-object]{RCX} or \link[=readCX]{CX} object}

\item{file}{character; the name of the file to which the data are written}

\item{verbose}{logical; whether to print what is happening}

\item{pretty}{logical; adds indentation whitespace to JSON output.
Can be TRUE/FALSE or a number specifying the number of spaces to indent. See \code{\link[jsonlite:prettify]{jsonlite::prettify()}}}
}
\description{
These function write an \link[=RCX-object]{RCX} object or a \link[=readCX]{CX} object to a file.
}
\examples{
NULL
}
\seealso{
\link{toCX}, \link{rcxToJson}, \link{readCX}
}
