% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AspectDependencies.R
\docType{data}
\name{aspectClasses}
\alias{aspectClasses}
\alias{subAspectClasses}
\alias{updateAspectClasses}
\title{aspectClasses and subAspectClasses}
\format{
An object of class \code{character} of length 14.

An object of class \code{character} of length 4.
}
\usage{
aspectClasses

subAspectClasses

updateAspectClasses(aspectClasses, aspect)
}
\arguments{
\item{aspectClasses}{named character; accession names and aspect classes}

\item{aspect}{named character; new accession names and aspect classes}
}
\value{
named character; accession names and aspect classes
}
\description{
\code{aspectClasses} and \code{subAspectClasses} contain the accession name and the classes of the corresponding (sub)aspect.
}
\details{
\code{updateAspectClasses} adds aspect classes to the list and avoids duplicate accession names and aspect classes.
}
\examples{
aspectClasses

aspectClasses = updateAspectClasses(
  aspectClasses, 
  c(bla="BlaAspect")
)

aspectClasses

aspectClasses = updateAspectClasses(
  aspectClasses, 
  c(bla="BlaAspect", blubb="BubbAspect")
)

aspectClasses
subAspectClasses
}
\keyword{datasets}
