% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-json.R
\name{convert2json}
\alias{convert2json}
\alias{.convert2json}
\alias{.convert2json.character}
\alias{.convert2json.numeric}
\alias{.convert2json.integer}
\alias{.convert2json.logical}
\alias{.convert2json.list}
\alias{.convert2json.data.frame}
\title{Convert data to json by R class}
\usage{
.convert2json(x, ...)

\method{.convert2json}{character}(x)

\method{.convert2json}{numeric}(x)

\method{.convert2json}{integer}(x)

\method{.convert2json}{logical}(x)

\method{.convert2json}{list}(x, raw = c(), byElement = FALSE, skipNa = TRUE)

\method{.convert2json}{data.frame}(x, raw = c(), skipNa = TRUE)
}
\arguments{
\item{x}{data element}

\item{raw}{character; names of columns not to format (e.g. because it is already converted)}
}
\value{
character; json
}
\description{
Convert data to json by R class
}
\note{
Internal function only for convenience
}
\examples{
NULL
}
\keyword{internal}
