% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-general.R
\name{.infer}
\alias{.infer}
\alias{.inferDataType}
\alias{.inferIsList}
\title{Infer the data type from values and check if the value elements are a list}
\usage{
.inferDataType(values)

.inferIsList(values)
}
\arguments{
\item{values}{vector or list of R data values}
}
\value{
character vector of data types or logical vector of list status
}
\description{
Each element has an R data type (i.e. class). If more than one element are present in one list element, it is marked as list
}
\details{
\code{.inferDataType} infers the data type of the elements in the vector or list.
\code{.inferIsList} infers for each element if it is a list. For a vector, the return therefore is \code{FALSE} for each element!
}
\section{Functions}{
\itemize{
\item \code{.inferIsList}: Infer, if the values are lists
}}

\note{
Internal function only for convenience
}
\examples{
NULL
}
\keyword{internal}
