% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCX.R
\docType{package}
\name{RCX}
\alias{RCX}
\title{R package implementing the Cytoscape Exchange (CX) format}
\description{
Create, handle, validate, visualize and convert networks in the Cytoscape exchange (CX) format to standard data types and objects.
}
\details{
The CX format is also used by the NDEx platform, a online commons for biological networks, and the network visualization software Cytocape.

\code{browseVignettes("RCy3")}
}
\author{
Florian Auer \email{florian.auer@informatik.uni-augsburg.de}
}
