% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{validate}
\alias{validate}
\alias{validate.default}
\alias{validate.NodesAspect}
\alias{validate.EdgesAspect}
\alias{validate.NodeAttributesAspect}
\alias{validate.EdgeAttributesAspect}
\alias{validate.NetworkAttributesAspect}
\alias{validate.CartesianLayoutAspect}
\alias{validate.CyGroupsAspect}
\alias{validate.CyVisualPropertiesAspect}
\alias{validate.CyVisualProperty}
\alias{validate.CyVisualPropertyProperties}
\alias{validate.CyVisualPropertyDependencies}
\alias{validate.CyVisualPropertyMappings}
\alias{validate.CyHiddenAttributesAspect}
\alias{validate.CyNetworkRelationsAspect}
\alias{validate.CySubNetworksAspect}
\alias{validate.CyTableColumnAspect}
\alias{validate.RCX}
\title{Validate RCX and its aspects}
\usage{
validate(x, verbose = T)

\method{validate}{default}(x, verbose = T)

\method{validate}{NodesAspect}(x, verbose = T)

\method{validate}{EdgesAspect}(x, verbose = T)

\method{validate}{NodeAttributesAspect}(x, verbose = T)

\method{validate}{EdgeAttributesAspect}(x, verbose = T)

\method{validate}{NetworkAttributesAspect}(x, verbose = T)

\method{validate}{CartesianLayoutAspect}(x, verbose = T)

\method{validate}{CyGroupsAspect}(x, verbose = T)

\method{validate}{CyVisualPropertiesAspect}(x, verbose = T)

\method{validate}{CyVisualProperty}(x, verbose = T)

\method{validate}{CyVisualPropertyProperties}(x, verbose = T)

\method{validate}{CyVisualPropertyDependencies}(x, verbose = T)

\method{validate}{CyVisualPropertyMappings}(x, verbose = T)

\method{validate}{CyHiddenAttributesAspect}(x, verbose = T)

\method{validate}{CyNetworkRelationsAspect}(x, verbose = T)

\method{validate}{CySubNetworksAspect}(x, verbose = T)

\method{validate}{CyTableColumnAspect}(x, verbose = T)

\method{validate}{RCX}(x, verbose = T)
}
\arguments{
\item{x}{object to validate; \link[=RCX-object]{RCX} object or an aspect}

\item{verbose}{logical; whether to print the test results.}
}
\value{
logical; whether the object passed all tests.
}
\description{
Validate RCX objects and its aspects.
}
\details{
Something.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default

\item \code{NodesAspect}: Nodes

\item \code{EdgesAspect}: Edges

\item \code{NodeAttributesAspect}: Node attributes

\item \code{EdgeAttributesAspect}: Edge attributes

\item \code{NetworkAttributesAspect}: Network attributes

\item \code{CartesianLayoutAspect}: Cartesian layout

\item \code{CyGroupsAspect}: Cytoscape Groups

\item \code{CyVisualPropertiesAspect}: Cytoscape Visual Properties

\item \code{CyVisualProperty}: Cytoscape Visual Properties

\item \code{CyVisualPropertyProperties}: Cytoscape visual property: Properties

\item \code{CyVisualPropertyDependencies}: Cytoscape visual property: Dependencies

\item \code{CyVisualPropertyMappings}: Cytoscape visual property: Mappings

\item \code{CyHiddenAttributesAspect}: Cytoscape hidden attributes

\item \code{CyNetworkRelationsAspect}: Cytoscape network relations

\item \code{CySubNetworksAspect}: Cytoscape sub-networks

\item \code{CyTableColumnAspect}: Cytoscape table column aspect

\item \code{RCX}: The whole RCX object with all its aspects
}}

\examples{
NULL
}
